Block placement  updated &  UI/sound inplemented.
Blocks are now placed using a better finding method, avoidning "jitter".  Blocks now check if the block being added and block it is adding to have suitable snapppoints using triggers. Placement UI sprite added to indicate when a block is allowed ot be palced or not (green for yes, red for no. It's even animated!). A builder "click" sound has been added upon sucessful addition. A new "not allowed to build there" sound will need to be added in next commit.