More water...
Yet more water changes. Removed the wave generator funciton as it was the reason water keeps disapearing. Will need to integrate gerstner waves, but this needs done in code as well as sahdergraph (and both must be identical, as code allows bouyoncy model to know where the water surface is, and shadergraph moves the visual representation of the surface)